Scheduling!
You can now set a SCHEDULE(* * *) in the Enabled column. This will only
run the test if it meets certain criteria.
Similar to a cron job you can set a pattern to control the days the test
should run.
day of the month (1-31)
|   month (1-12 or JAN-DEC
|  |  day of the week (0-6 where Sunday is 0 or SUN-SAT)
|  |  |
SCHEDULE(* * *)

e.g SCHEDULE(1-7 * MON)   Will cause the test to run on the first Monday
of the month.
SCHEDULE(* DEC *) Will cause the test to only run in December
SCHEDULE(* * TUE,THU,FRI-SUN) Will mean the test only runs on Tuesdays,
thursdays and friday through to sunday.